backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
pydantic>=2.0.0
opencv-python-headless>=4.9.0
mediapipe>=0.10.0
numpy>=1.24.0
